+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to the "align-spaces" extension will be documented in this f
 ## PATCH
 
 -   Better handling for multi-character tokens
+-   Don't align lines with different indentation
+-   Don't align lines with different types of operators (eg. assignment `a = 1;` with comparison `if (a == 1)`)
+-   Group sets of object property assignments
 
 ## 0.0.1